I think I have the rest of the git stuff done
Needs testing

+++ app.rb
@@ -10,7 +10,7 @@ require_relative 'public_app'
 
 module AwestructWebEditor
   class SecureApp < Sinatra::Base
-    #set :ssl, lambda { !development? }
+    set :ssl, lambda { |_| !development? }
     register Sinatra::Sprockets::Helpers
 
     use AwestructWebEditor::PublicApp
@@ -20,7 +20,8 @@ module AwestructWebEditor
       require 'sinatra/reloader'
       register Sinatra::Reloader
       also_reload 'app.rb'
-      also_reload 'models/**/*.rb'
+      also_reload 'helpers/**/*.rb'
+      also_reload 'public_app.rb'
       set :raise_errors, true
       enable :logging, :dump_errors, :raise_errors
     end
@@ -44,8 +45,12 @@ module AwestructWebEditor
       end
 
       def read_settings
-        File.open(settings_storage_file, 'r') do |f|
-          JSON.load(f)
+        if File.exists?(settings_storage_file)
+          File.open(settings_storage_file, 'r') do |f|
+            JSON.load(f)
+          end
+        else
+          ''
         end
       end
